+++ cicero/pkg/tcp/routes/cal.go
@@ -11,11 +11,10 @@ import (
 
 	"github.com/zimeg/quintus/cicero/pkg/cal"
 	"github.com/zimeg/quintus/cicero/pkg/now"
-	"github.com/zimeg/quintus/cicero/pkg/utc"
 )
 
 // calendar formats a table of dates to output
-func calendar(year int) (w bytes.Buffer, err error) {
+func calendar(year int, today now.Now) (w bytes.Buffer, err error) {
 	funcs := template.FuncMap{
 		"format": func(t time.Time) string {
 			return t.Format("2006-01-02")
@@ -24,10 +23,9 @@ func calendar(year int) (w bytes.Buffer, err error) {
 			return i % j
 		},
 		"today": func(date now.Now) bool {
-			current := now.Moment(utc.Now())
-			return date.Year() == current.Year() &&
-				date.Month() == current.Month() &&
-				date.Date() == current.Date()
+			return date.Year() == today.Year() &&
+				date.Month() == today.Month() &&
+				date.Date() == today.Date()
 		},
 	}
 	tpl := `
@@ -35,30 +33,37 @@ func calendar(year int) (w bytes.Buffer, err error) {
 {{- $d := index $month 0 -}}
 {{- $id := printf "%d-%02d" $d.Quintus.Year $d.Quintus.Month }}
 <tr>
-    <th id="{{ $id }}" colspan="6">
-        <a href="#{{ $id }}">{{ $id }}</a>
-    </th>
+	<th id="{{ $id }}" colspan="6">
+		<a href="#{{ $id }}">{{ $id }}</a>
+	</th>
 </tr>
 <tr>
-    {{- range $num, $dates := $month }}
-    <td title="{{ format $dates.Gregorian }}">
-        <label>
-        <p>
-            <input type="radio"
-            {{ if (today $dates.Quintus) }}checked{{ end }}
-                name="date"
-                hx-get="/date/{{ format $dates.Gregorian }}"
-                hx-target="#timers"
-                hx-swap="outerHTML">
-            {{ if (lt $dates.Quintus.Date 10) }}&nbsp;{{ end }}{{ $dates.Quintus.Date }}
-        </p>
-        </label>
-    </td>
-    {{- if and (ne $dates.Quintus.Month 0) (eq (mod $dates.Quintus.Date 5) 0) (ne $dates.Quintus.Date 30) }}
+	{{- range $num, $dates := $month }}
+	<td
+		title="{{ format $dates.Gregorian }}"
+		id="{{ printf "Q%d-%02d-%02d" $dates.Quintus.Year $dates.Quintus.Month $dates.Quintus.Date }}"
+	>
+		<label>
+			<p>
+				<input
+					type="radio"
+					{{ if (today $dates.Quintus) }}checked{{ end }}
+					name="date"
+					value="{{ format $dates.Gregorian }}"
+					hx-get="/date/{{ format $dates.Gregorian }}"
+					hx-params="none"
+					hx-target="#timers"
+					hx-swap="outerHTML"
+				>
+				{{ if (lt $dates.Quintus.Date 10) }}&nbsp;{{ end }}{{ $dates.Quintus.Date }}
+			</p>
+		</label>
+	</td>
+	{{- if and (ne $dates.Quintus.Month 0) (eq (mod $dates.Quintus.Date 5) 0) (ne $dates.Quintus.Date 30) }}
 </tr>
 <tr>
-    {{- end }}
-    {{- end }}
+	{{- end }}
+	{{- end }}
 </tr>
 {{- end}}`
 	t, err := template.New("cal").Funcs(funcs).Parse(tpl)
@@ -96,7 +101,7 @@ func Cal(w http.ResponseWriter, r *http.Request) {
 		fmt.Fprintf(w, "%d-04-00T03:55:05Z", current.Year())
 		return
 	}
-	cal, err := calendar(year)
+	cal, err := calendar(year, current)
 	if err != nil {
 		w.WriteHeader(http.StatusInternalServerError)
 		fmt.Fprintf(w, "%d-05-00T03:55:05Z", current.Year())
@@ -108,10 +113,10 @@ func Cal(w http.ResponseWriter, r *http.Request) {
 <tr id="before"></tr>
 %s
 <tr hx-get="/cal/%d"
-  hx-disable
-  hx-target="#before"
-  hx-trigger="revealed once"
-  hx-swap="outerHTML show:[id='%d-00']:top"
+	hx-disable
+	hx-target="#before"
+	hx-trigger="revealed once"
+	hx-swap="outerHTML show:[id='%d-00']:top"
 >
 </tr>`,
 			strings.TrimSpace(cal.String()),
@@ -121,10 +126,10 @@ func Cal(w http.ResponseWriter, r *http.Request) {
 	case year > current.Year():
 		fmt.Fprintf(w, `
 <tr hx-get="/cal/%d"
-  hx-disable
-  hx-target="#after"
-  hx-trigger="revealed once"
-  hx-swap="outerHTML"
+	hx-disable
+	hx-target="#after"
+	hx-trigger="revealed once"
+	hx-swap="outerHTML"
 >
 %s
 <tr id="after"></tr>