+++ cicero/pkg/tcp/routes/date.go
@@ -11,9 +11,18 @@ import (
 
 // Date returns timers for the provided Gregorian date
 func Date(w http.ResponseWriter, r *http.Request) {
-	current := now.Moment(utc.Now())
+	cookie, _ := r.Cookie("timezone")
+	timezone := "Etc/UTC"
+	if cookie != nil {
+		timezone = cookie.Value
+	}
+	location, err := time.LoadLocation(timezone)
+	if err != nil {
+		location = time.UTC
+	}
+	current := now.Moment(utc.Current().In(location).ToTime())
 	query := r.PathValue("date")
-	gregorian, err := time.Parse("2006-01-02", query)
+	gregorian, err := time.ParseInLocation("2006-01-02", query, location)
 	if err != nil {
 		w.WriteHeader(http.StatusBadRequest)
 		fmt.Fprintf(w, "%d-04-00T03:55:05Z", current.Year())
@@ -23,27 +32,51 @@ func Date(w http.ResponseWriter, r *http.Request) {
 	if quintus.Year() == current.Year() &&
 		quintus.Month() == current.Month() &&
 		quintus.Date() == current.Date() {
-		fmt.Fprintf(w, `<article id="timers">
-			<p>
-				<span title="Quintus Time Server">QTS</span>
+		fmt.Fprintf(w, `
+			<article
+				id="timers"
+				hx-post="/timezone"
+				hx-swap="outerHTML"
+				hx-vals="js:{timezone: Intl.DateTimeFormat().resolvedOptions().timeZone}"
+			>
+				<p>
+					<span title="Quintus Time Server">QTS</span>
 					......
-					<time id="quintus" hx-get="/now" hx-trigger="every 1s">
+					<time
+						id="quintus"
+						hx-get="/now"
+						hx-params="none"
+						hx-swap="innerHTML"
+						hx-trigger="every 1s"
+					>
 						%s
 					</time>
 				</p>
 				<p>
 					<span title="Universal Coordinated Time">UTC</span>
 					......
-					<time id="utc" hx-get="/utc" hx-trigger="every 1s">
+					<time
+						id="utc"
+						hx-get="/utc"
+						hx-params="none"
+						hx-trigger="every 1s"
+						hx-swap="innerHTML"
+					>
 						%s
 					</time>
 				</p>
-			</article>`, current.ToString(), utc.ToString())
+			</article>`, current.ToString(), utc.Current().In(location).ToString())
 		return
 	}
-	fmt.Fprintf(w, `<article id="timers">
-		<p>
-			<span title="Quintus Time Server">QTS</span>
+	fmt.Fprintf(w, `
+		<article 
+			id="timers"
+			hx-post="/timezone"
+			hx-swap="outerHTML"
+			hx-vals="js:{timezone: Intl.DateTimeFormat().resolvedOptions().timeZone}"
+		>
+			<p>
+				<span title="Quintus Time Server">QTS</span>
 				......
 				<time id="quintus">
 					%s
@@ -56,5 +89,5 @@ func Date(w http.ResponseWriter, r *http.Request) {
 					%s
 				</time>
 			</p>
-		</article>`, quintus.ToString(), gregorian.Format(time.RFC3339))
+		</article>`, quintus.ToString(), utc.Moment(gregorian).ToString())
 }