+++ CHANGELOG.md
@@ -8,6 +8,7 @@ calendar.
 
 ## Changes
 
+- feat: show a date for the timezone sent in requests to time server 2025-12-01
 - feat: include metadata meant to find these pages from between site 2025-11-17
 - feat: link to a checkout page for selling calendars of soon months 2025-11-12
 - fix: inline radio buttons on mobile devices for a dense conversion 2025-10-16